l10n_es_facturae: Code organization + refactoring
