+++ l10n_es_facturae/__openerp__.py
@@ -1,24 +1,11 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    OpenERP, Open Source Management Solution
-#    Copyright (c) 2009 Alejandro Sanchez (http://www.asr-oss.com)
-#                       Alejandro Sanchez <alejandro@asr-oss.com>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the Affero GNU General Public License as published
-#    by the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the Affero GNU General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# © 2009 Alejandro Sanchez <alejandro@asr-oss.com>
+# © 2015 Ismael Calvo <ismael.calvo@factorlibre.com>
+# © 2015 Tecon
+# © 2015 Juanjo Algaz (MalagaTIC)
+# © 2015 Omar Castiñeira (Comunitea)
+# © 2016 Serv. Tecnol. Avanzados - Pedro M. Baeza
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 {
     "name": "Creación de Factura-e (FACe)",
@@ -39,9 +26,9 @@
         "base_iso3166",
     ],
     "data": [
-        "partner_view.xml",
-        "res_company.xml",
-        "payment_mode_view.xml",
+        "views/res_partner_view.xml",
+        "views/res_company.xml",
+        "views/payment_mode_view.xml",
         "wizard/create_facturae_view.xml",
         "data/l10n_es_facturae_data.xml",
         "security/ir.model.access.csv",
