+++ l10n_es_facturae/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from . import models
-from . import wizard
+from . import res_partner
+from . import res_company
+from . import payment_mode
