+++ l10n_es_facturae/views/payment_mode_view.xml
