+++ l10n_es_facturae/wizard/create_facturae.py
@@ -1,26 +1,11 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    OpenERP, Open Source Management Solution
-#    Copyright (c) 2009 Alejandro Sanchez (http://www.asr-oss.com)
-#    Alejandro Sanchez <alejandro@asr-oss.com>
-#    Copyright (c) 2015 Comunitea Servicios Tecnológicos
-#    Omar Castiñeira Saavedra <omar@comunitea.com>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the Affero GNU General Public License as published
-#    by the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the Affero GNU General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# © 2009 Alejandro Sanchez <alejandro@asr-oss.com>
+# © 2015 Ismael Calvo <ismael.calvo@factorlibre.com>
+# © 2015 Tecon
+# © 2015 Juanjo Algaz (MalagaTIC)
+# © 2015 Omar Castiñeira (Comunitea)
+# © 2016 Serv. Tecnol. Avanzados - Pedro M. Baeza
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 import base64
 from openerp import models, fields, _, api, exceptions
@@ -266,38 +251,37 @@ class CreateFacturae(models.TransientModel):
             def create_administrative_centres(address, code):
                 texto = ""
                 if administrative:
+                    texto += "<AdministrativeCentre>"
                     if code == '01':
-                        texto += "<AdministrativeCentre>"
-                        texto += "<CentreCode>" + address.oficina_contable[:10] +\
-                                 "</CentreCode>"
-                        texto += "<RoleTypeCode>" + code + "</RoleTypeCode>"
+                        texto += (
+                            "<CentreCode>" + address.oficina_contable[:10] +
+                            "</CentreCode>")
                     elif code == '02':
-                        texto += "<AdministrativeCentre>"
-                        texto += "<CentreCode>" + address.organo_gestor[:20] +\
-                                 "</CentreCode>"
-                        texto += "<RoleTypeCode>" + code + "</RoleTypeCode>"
+                        texto += (
+                            "<CentreCode>" + address.organo_gestor[:20] +
+                            "</CentreCode>")
                     elif code == '03':
-                        texto += "<AdministrativeCentre>"
-                        texto += "<CentreCode>" + address.unidad_tramitadora[:20] +\
-                                 "</CentreCode>"
-                        texto += "<RoleTypeCode>" + code + "</RoleTypeCode>"
+                        texto += (
+                            "<CentreCode>" + address.unidad_tramitadora[:20] +
+                            "</CentreCode>")
                     elif code == '04':
-                        texto += "<AdministrativeCentre>"
-                        texto += "<CentreCode>" + address.organo_proponente[:20] +\
-                                 "</CentreCode>"
-                        texto += "<RoleTypeCode>" + code + "</RoleTypeCode>"
+                        texto += (
+                            "<CentreCode>" + address.organo_proponente[:20] +
+                            "</CentreCode>")
                     else:
-                        texto += "<AdministrativeCentre>"
                         texto += "<CentreCode>" + '' + "</CentreCode>"
-                        texto += "<RoleTypeCode>" + code + "</RoleTypeCode>"
+                    texto += "<RoleTypeCode>" + code + "</RoleTypeCode>"
 
                 texto += '<AddressInSpain>'
                 if address.street:
                     if address.street2:
-                        texto += '<Address>' + (address.street + ' ' + \
-                                 address.street2)[:80] + '</Address>'
+                        texto += (
+                            '<Address>' + (address.street + ' ' +
+                                           address.street2)[:80] +
+                            '</Address>')
                     else:
-                        texto += '<Address>' + address.street[:80] + '</Address>'
+                        texto += (
+                            '<Address>' + address.street[:80] + '</Address>')
                 else:
                     log.add(_('User error:\n\nPartner %s has no street.') %
                             address.name, True)
@@ -340,7 +324,9 @@ class CreateFacturae(models.TransientModel):
                 if address.fax:
                     texto += '<TeleFax>' + address.fax[:15] + '</TeleFax>'
                 if address.website:
-                    texto += '<WebAddress>' + address.website[:60] + '</WebAddress>'
+                    texto += (
+                        '<WebAddress>' + address.website[:60] +
+                        '</WebAddress>')
                 if address.email:
                     texto += '<ElectronicMail>' + address.email[:60] +\
                              '</ElectronicMail>'
