+++ l10n_es_facturae/wizard/create_facturae_view.xml
@@ -19,20 +19,28 @@
                         <field name="note" colspan="4" nolabel="1"/>
                     </group>
                     <footer>
-                        <button icon="gtk-print" name="create_facturae_file" string="Exportar" type="object" states="first" class="oe_highlight"/>
-                        o
-                        <button string="Cerrar" class="oe_link" special="cancel"/>
+                        <button icon="gtk-print"
+                                name="create_facturae_file"
+                                string="Export"
+                                type="object"
+                                states="first"
+                                class="oe_highlight"/>
+                        or
+                        <button string="Close"
+                                class="oe_link"
+                                special="cancel"/>
                     </footer>
                 </form>
             </field>
         </record>
 
         <act_window name="Create Factura-e file"
-            res_model="create.facturae"
-            src_model="account.invoice"
-            view_mode="form"
-            target="new"
-            id="wizard_create_facturae_act_window" />
+                    res_model="create.facturae"
+                    src_model="account.invoice"
+                    view_mode="form"
+                    target="new"
+                    id="wizard_create_facturae_act_window"
+        />
 
     </data>
 </openerp>
